Fix: Tooltip of the file name tab was not updated if renamed Fix: Use min size of the config and actual resolution

+++ src/MadEditApp.cpp
@@ -489,8 +489,8 @@ bool MadEditApp::OnInit()
 		if( x + w > 0 && y + h > 0 )
 			//if(w>0 && h>0)
 		{
-			size.x = std::min( w, size.x );
-			size.y = std::min( h, size.y );
+			size.x = std::min( (int)w, size.x );
+			size.y = std::min( (int)h, size.y );
 			pos.x = x;
 			pos.y = y;
 		}
